+++ markdown/__tests__/setup.js
@@ -28,12 +28,7 @@ import {
 
 import { markdownParser } from '../markdown-parser';
 import { SpecRegistry } from '@bangle.dev/core/spec-registry';
-import {
-  table,
-  tableCell,
-  tableHeader,
-  tableRow,
-} from '@bangle.dev/core/components/table/index';
+import { table, tableCell, tableHeader, tableRow } from '@bangle.dev/table';
 
 const specRegistry = new SpecRegistry([
   // nodes
@@ -77,4 +72,6 @@ export const serialize = async (doc) => {
   return serializer.serialize(content);
 };
 
-export const parse = async (md) => parser.parse(md);
+export const parse = async (md) => {
+  return parser.parse(md);
+};